YOLO-OCR-NLP/requirements.txt
numpy==1.26.4
scipy==1.13.1
scikit-image==0.22.0
Pillow==10.0.0
matplotlib==3.8.0
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
shapely==2.0.2
albumentations==1.3.1
imgaug==0.4.0
paddleocr==2.7.0.3
beautifulsoup4
pyclipper
lmdb
tqdm
visualdl
rapidfuzz
cython
lxml
premailer
openpyxl
attrdict
pyyaml
python-docx
fonttools
fire
ultralytics
